backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Observability / metrics
prometheus-client==0.21.0

# Settings
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic[email]

# Database (async)
sqlalchemy==2.0.36
asyncpg==0.30.0

# Redis (async)
redis==5.2.0

# Auth / security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1              # pinned: passlib 1.7.4 is incompatible with bcrypt>=4.1
cryptography>=42.0.0       # Fernet field-level encryption

# Rate limiting
slowapi==0.1.9

# Ingestion / job queue
celery==5.4.0
python-multipart==0.0.12

# FHIR R4 HTTP client (httpx already included below)
httpx==0.27.2

# CSV / Excel parsing
openpyxl==3.1.5

# Vision / Image processing
Pillow>=10.0.0

# ML / Forecasting
prophet==1.1.5
lightgbm==4.5.0
pandas==2.2.3
numpy==1.26.4
mlflow==2.17.2
scikit-learn==1.5.2

# Vision ML (VVM classifier)
torch>=2.2.0
onnx>=1.14.0
onnxruntime>=1.16.0

# Date arithmetic (also pulled in by pandas/prophet, listed explicitly for scripts)
python-dateutil>=2.9.0

# Dev / test
pytest==8.3.3
pytest-asyncio==0.24.0

# Database migrations
alembic==1.13.3

# Object detection (AR stock counter)
ultralytics>=8.0.0
scipy>=1.10.0
